+++ Public/Get-Warranty.ps1
@@ -22,7 +22,7 @@ function Get-Warranty {
             # RMM Mode
             [Parameter(Mandatory = $false)]
             [ValidateSet('NinjaRMM', 'None')]
-		    [String]$RMM = 'NinjaRMM',
+            [String]$RMM = 'NinjaRMM',
             #Enable Registry Storing
             [Parameter(Mandatory = $false)]
             [bool]$EnableRegistry = $true,
@@ -30,7 +30,7 @@ function Get-Warranty {
             [String]$RegistryPath = 'HKLM:\SOFTWARE\RMMCustomInfo\',
             # Force Update RMM with details
             [Parameter(Mandatory = $false)]
-		    [bool]$ForceUpdate = $false,
+            [bool]$ForceUpdate = $false,
             # Custom Machine Details
             [Parameter(Mandatory = $false)]
             [String]$Serial = 'Automatic',
@@ -72,7 +72,7 @@ function Get-Warranty {
             $mfg = $Manufacturer
         }
         
-
+        $Notsupported = $false
         switch -Wildcard ($mfg){
             "EDSYS"{
                 $Warobj = Get-WarrantyEdsys -Serial $serialnumber -DateFormat $DateFormat
@@ -89,16 +89,60 @@ function Get-Warranty {
             "HP"{
                 $Warobj = Get-WarrantyHP -Serial $serialnumber -DateFormat $DateFormat
             }
+            "MICROSOFT"{
+                if($($machineinfo.Model) -like 'SurfaceNotSupportedYet'){
+                    $Warobj = Get-WarrantyMicrosoft -Serial $serialnumber -DateFormat $DateFormat
+                } else{
+                    $Notsupported = $true
+                    Write-Host "Microsoft Model not Supported"
+                    Write-Host "Manufacturer  :  $mfg"
+                    Write-Host "Model         :  $($machineinfo.Model)"
+                }
+                
+            }
+            "TOSHIBA"{
+                $Warobj = Get-WarrantyToshiba -Serial $serialnumber -DateFormat $DateFormat
+            }
             default{
                 $Notsupported = $true
-                Write-Host "Manufacturer not Supported :  $mfg"
+                Write-Host "Manufacturer or Model not Supported"
+                Write-Host "Manufacturer  :  $mfg"
+                Write-Host "Model         :  $($machineinfo.Model)"
             }
         }
     if($RMM -eq 'NinjaRMM' -and ($Notsupported -eq $false)){
-        Write-WarrantyNinjaRMM -DateFormat $DateFormat -Warrantystart $($WarObj.'StartDate') -WarrantyExpiry $($WarObj.'EndDate') -WarrantyStatus $($WarObj.'Warranty Status') -Invoicenumber $($WarObj.'Invoice')
+        $ParamsNinjaRMM = @{
+            DateFormat = $DateFormat
+        }
+        if ($WarObj.'StartDate') {
+            $ParamsNinjaRMM['Warrantystart'] = $WarObj.'StartDate'
+        }
+        if ($WarObj.'EndDate') {
+            $ParamsNinjaRMM['WarrantyExpiry'] = $WarObj.'EndDate'
+        }
+        if ($WarObj.'Warranty Status') {
+            $ParamsNinjaRMM['WarrantyStatus'] = $WarObj.'Warranty Status'
+        }
+        if ($WarObj.'Invoice') {
+            $ParamsNinjaRMM['Invoicenumber'] = $WarObj.'Invoice'
+        }
+        Write-WarrantyNinjaRMM @ParamsNinjaRMM
     }
     if($EnableRegistry -and ($Notsupported -eq $false)){
-        Write-WarrantyRegistry -RegistryPath $RegistryPath -Warrantystart $($WarObj.'StartDate') -WarrantyExpiry $($WarObj.'EndDate') -WarrantyStatus $($WarObj.'Warranty Status') -Invoicenumber $($WarObj.'Invoice')
+        $Params = @{}
+        if ($WarObj.'StartDate') {
+            $Params['Warrantystart'] = $WarObj.'StartDate'
+        }
+        if ($WarObj.'EndDate') {
+            $Params['WarrantyExpiry'] = $WarObj.'EndDate'
+        }
+        if ($WarObj.'Warranty Status') {
+            $Params['WarrantyStatus'] = $WarObj.'Warranty Status'
+        }
+        if ($WarObj.'Invoice') {
+            $Params['Invoicenumber'] = $WarObj.'Invoice'
+        }
+        Write-WarrantyRegistry -RegistryPath $RegistryPath @Params
     }
 return $Warobj
 }